restructure go utils directory
