+++ go-utils/abi-bindings/CrossChainApplications/ERC20Bridge/BridgeToken/BridgeToken.go
